obstacle_sensor field became nearby_vessel_poses
Replaced binary "obstacle_sensor" with artifically limited communication
range data. If another musv is within a certain distance of this vessel,
It is assumed they are able to communicate their poses to each other.
This communication is represented by the inclusion of the poses of all
other vessels within the threshold distance in the sensorData message.